tweaks & improvements + fedora almost supported

+++ distros/ubuntu.sh
@@ -3,7 +3,16 @@ PM_UPGRADE="apt full-upgrade -y"
 PM_INSTALL="apt install -y"
 PACKAGES="curl git zsh vlc ffmpeg flatpak clang fonts-firacode python3-pip"
 
-# VSCode
-curl -L "https://code.visualstudio.com/sha/download?build=stable&os=linux-deb-x64" > vscode.deb
-sudo $PM_INSTALL ./vscode.deb
-rm vscode.deb
+setup() {
+	# VSCode
+	if ! command -v code; then
+		curl -L "https://code.visualstudio.com/sha/download?build=stable&os=linux-deb-x64" > vscode.deb
+		sudo $PM_INSTALL ./vscode.deb
+		rm vscode.deb
+	fi
+}
+
+clean() {
+	sudo apt autoremove -y
+	sudo apt clean
+}
